+++ federation/mod.ts
@@ -11,5 +11,7 @@ export {
   respondWithObjectIfAcceptable,
   type RespondWithObjectOptions,
 } from "./handler.ts";
+export * from "./kv.ts";
 export * from "./middleware.ts";
+export * from "./mq.ts";
 export * from "./router.ts";